refactor: prefer put list updater functions
This commit refactors the functionality for adding items, deleting
individual items, and deleting the last item from a shopping list. The
goal of this refactor is to make the functions that update a shopping
list in some way more "single responsibility". The previous
implementation tied the various pieces of list update behavior to the
contextual state of the `ShoppingList` component. However, the actual
"state" of the shopping list is not related the pure functionality of
adding to or removing items from a list of "shopping items".

These changes include some fundamental Functional Programming techniques
like:

1. Currying - https://medium.com/javascript-scene/curry-and-function-composition-2c208d774983
2. Point free style - https://medium.com/dailyjs/functional-js-7-point-free-style-b21a1416ac6a